+++ judgels-backends/sandalphon/sandalphon-client/src/main/java/judgels/sandalphon/submission/AbstractSubmissionClient.java
@@ -8,8 +8,8 @@ import java.io.IOException;
 import judgels.gabriel.api.GradingRequest;
 import judgels.gabriel.api.SubmissionSource;
 import judgels.sandalphon.api.problem.programming.ProblemSubmissionConfig;
-import judgels.sandalphon.api.submission.ProgrammingSubmission;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionData;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmission;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionData;
 import judgels.sandalphon.persistence.AbstractGradingModel;
 import judgels.sandalphon.persistence.AbstractSubmissionModel;
 import judgels.sealtiel.api.message.MessageData;
