+++ judgels-backends/sandalphon/sandalphon-client/src/main/java/judgels/sandalphon/submission/AbstractSubmissionStore.java
@@ -18,8 +18,8 @@ import judgels.persistence.api.OrderDir;
 import judgels.persistence.api.Page;
 import judgels.persistence.api.SelectionOptions;
 import judgels.sandalphon.api.submission.Grading;
-import judgels.sandalphon.api.submission.ProgrammingSubmission;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionData;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmission;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionData;
 import judgels.sandalphon.persistence.AbstractGradingModel;
 import judgels.sandalphon.persistence.AbstractSubmissionModel;
 import judgels.sandalphon.persistence.BaseGradingDao;
