+++ judgels-backends/uriel/uriel-app/src/integTest/java/judgels/uriel/contest/submission/ContestProgrammingSubmissionStoreIntegrationTests.java
@@ -8,8 +8,8 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import judgels.persistence.TestActorProvider;
 import judgels.persistence.hibernate.WithHibernateSession;
-import judgels.sandalphon.api.submission.ProgrammingSubmission;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionData;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmission;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionData;
 import judgels.uriel.AbstractIntegrationTests;
 import judgels.uriel.UrielIntegrationTestComponent;
 import judgels.uriel.api.contest.Contest;
