+++ judgels-backends/uriel/uriel-app/src/main/java/judgels/uriel/contest/scoreboard/ioi/IoiScoreboardProcessor.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 import judgels.gabriel.api.Verdict;
 import judgels.gabriel.api.Verdicts;
 import judgels.sandalphon.api.submission.Grading;
-import judgels.sandalphon.api.submission.ProgrammingSubmission;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmission;
 import judgels.uriel.api.contest.Contest;
 import judgels.uriel.api.contest.module.IoiStyleModuleConfig;
 import judgels.uriel.api.contest.module.StyleModuleConfig;
