+++ judgels-backends/uriel/uriel-app/src/main/java/judgels/uriel/contest/submission/programming/ContestProgrammingSubmissionResource.java
@@ -37,10 +37,10 @@ import judgels.sandalphon.SandalphonUtils;
 import judgels.sandalphon.api.client.problem.ClientProblemService;
 import judgels.sandalphon.api.problem.ProblemInfo;
 import judgels.sandalphon.api.problem.programming.ProblemSubmissionConfig;
-import judgels.sandalphon.api.submission.ProgrammingSubmission;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionData;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionWithSource;
-import judgels.sandalphon.api.submission.ProgrammingSubmissionWithSourceResponse;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmission;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionData;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionWithSource;
+import judgels.sandalphon.api.submission.programming.ProgrammingSubmissionWithSourceResponse;
 import judgels.service.actor.ActorChecker;
 import judgels.service.api.actor.AuthHeader;
 import judgels.service.api.client.BasicAuthHeader;
